requirements.txt
pytest==5.4.1
py-solc-x==0.1.1
web3==4.8.3
eth-brownie==1.1.0
